+++ src/myTokens/token2.sol
@@ -5,11 +5,9 @@ import "@openzeppelin/ERC20/IERC20.sol";
 import "@openzeppelin/ERC20/extensions/ERC20Permit.sol";
 
 contract Token2 is ERC20 {
-    constructor() ERC20("Token2", "T2") {
-     
-    }
+    constructor() ERC20("Token2", "T2") {}
 
-    function mint(address add, uint amount) public {
-        _mint(add , amount);
+    function mint(address add, uint256 amount) public {
+        _mint(add, amount);
     }
 }
